+++ src/client/components/selectable-string-filter-menu/selectable-string-filter-menu.scss
@@ -1,0 +1,106 @@
+/*
+ * Copyright 2015-2016 Imply Data, Inc.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+@import '../../imports';
+
+$checkbox-width: 20px;
+
+.string-filter-menu {
+  .menu-table {
+    .side-by-side {
+      @extend %side-by-side;
+
+      :not(:last-child) {
+        margin-right: 0;
+      }
+    }
+
+    @extend %white-shadow;
+    background: $white;
+    color: $text-standard;
+
+    &::after {
+      @include unpin-full(0, -14px, 0, -14px);
+    }
+
+    .search-box {
+      height: 30px;
+      margin-left: 6px;
+
+      .clearable-input {
+        @extend %input-cont-inset;
+        width: 100%;
+        height: $input-height;
+        padding-left: 6px;
+      }
+    }
+
+    .rows {
+      height: 280px;
+      margin-top: 10px;
+      margin-left: -$padding-compact;
+      margin-right: -$padding-compact;
+
+      overflow: auto;
+    }
+
+    .row {
+      height: 24px;
+      cursor: pointer;
+      padding: 0 $padding-compact;
+
+      .row-wrapper {
+        @include ellipsis;
+        height: 100%;
+
+        .label {
+          display: inline-block;
+          vertical-align: top;
+          padding-top: 4px;
+        }
+      }
+
+      &:hover {
+        background: $hover;
+      }
+
+      &:last-child {
+        margin-bottom: 12px;
+      }
+
+      &.no-select {
+        pointer-events: none;
+        user-select: none;
+      }
+    }
+
+    .message {
+      padding: 4px 10px;
+      color: $text-light;
+      font-style: italic;
+    }
+
+    .matching-values-message {
+      padding: 4px 10px;
+      color: $text-light;
+    }
+
+    .loader,
+    .query-error {
+      @include pin-full;
+    }
+  }
+}
