+++ src/client/components/string-filter-menu-controls/string-filter-menu-controls.scss
@@ -1,0 +1,24 @@
+@import '../../imports';
+
+.string-filter-menu-controls {
+  .side-by-side {
+    @extend %side-by-side;
+
+    :not(:last-child) {
+      margin-right: 0;
+    }
+  }
+
+  .search-box {
+    height: 30px;
+    margin-left: 6px;
+    width: 100%;
+
+    .clearable-input {
+      @extend %input-cont-inset;
+      width: 100%;
+      height: $input-height;
+      padding-left: 6px;
+    }
+  }
+}
